Removing all memory allocation from audio thread
This change introduces a new and different way of handling callbacks.
It removes much work from the audio thread, including the conversion
from plain C datatypes to Objective-C datatypes.  Instead, the audio
thread writes the parameters of callbacks to a lockless circular buffer,
and an intermediate converter thread reads the circular buffer, converts
to Objective-C datatypes, and puts the callbacks on the main queue.

WARNING:  This branch is very experimental.  It may disappear without
warning or blow up in your face.  Use at your own risk.