Update PdAudioController.m
Ok, iOS 6 now handles Playback category as well. It actually wasn't handled at all before. Unfortunately, I don't have the ability to test on iOS 6 since Xcode 8 is 8.0+, but I think it looks good. Would be great if someone could test though. Or perhaps it may be time to support only 8.0 and up as well?

One thing to note, as mentioned in the TODO above, for the Playback category, the MixWithOthers property is currently set even if the allowsMixing userFlag is set to NO.  Handling this properly requires an update to how PdAudioController handles interruptions. Namely, endInterruptionWithFlags: needs to updated since the ShouldResume flag has been deprecated. 

All the best,
Alexander